Use snapshot data in WindowsFormsEditorProvider
Fixes: #6726

Under high project lock contention, such as during a branch switch or a change to a props/targets shared by an entire solution, we can hit UI delays selecting items in Solution Explorer trying to figure out if we can "preview" files.

Use snapshot data if available in WindowsFormsEditorProvider to prevent that.