Use TargetFramework instead of TargetFrameworkMoniker
This fixes an issue where the "Target framework" drop down didn't show the current value.

The drop down is getting the right set of _supported_ values, and if you expand the drop down you will see the display names ".NET Core 2.0", ".NET Core 2.1", ".NET Core 3.0", etc. These display names and the corresponding values (.netcoreapp2.0, .netcoreapp2.1, .netcoreapp3.0, etc.) are ultimately coming from the `SupportedTargetFrameworksProvider`.

However, the _current_ value is coming from the "TargetFrameworkMoniker" property, and that value has the form ".NETCoreApp,Version=3.0" or similar. So the current value doesn't match any of the admissible values.

The solution here is to bind directly to the "TargetFramework" property instead of "TargetFrameworkMoniker". The older property pages use "TargetFrameworkMoniker" due to the requirements of the CSProj project system, and it was carried over into the new pages.