Support a custom manifest file path
The Application property page .xaml file currently has an "ApplicationManifest" property that is backed by the `ApplicationManifestValueProvider` interceptor. The interceptor supports three values: "NoManifest", indicating that the application should not have a manifest at all; "DefaultManifest", indicating that the build should provide a default manifest; or a path to a custom manifest file. Depending on which of these is provided, the interceptor sets the "ApplicationManifest" and/or "NoWin32Manifest" MSBuild properties in the project file.

In the legacy property pages, this is represented with a combo box that lets the user select "Create application without a manifest" or "Embed manifest with default settings" from the drop down, or type in the file path. In the new property pages, however, we don't currently have support for such a combo box. This means there is currently no way to enter the path to a custom manifest file.

This change splits the functionality into two properties in the .xaml file, "ApplicationManifestKind" to represent no manifest, default manifest, or custom manifest, and "ApplicationManifestKind", strictly to hold the value of the custom manifest path when one is set. The latter is only visible when the manifest kind is set to "CustomManifest".

We still need the `ApplicationManifestValueProvider` to back the property in the legacy property pages, so these new properties are backed by new interceptors, `ApplicationManifestKindValueProvider` and `ApplicationManifestPathValueProvider`. These largely borrow code from `ApplicationManifestValueProvider`.