Temporarily store cleared values
The `ITemporaryPropertyStorage` interface provides for the temporary storage of configuration-independent property values. Consider the project property pages. In some cases, changing the value of property A will cause us to clear the value of property B. If A is changed back to its original value, we would like to be able to restore the previous value of B as well. This type provides the means to store that previous value until (and if) we need it again.

This change introduces the interface and implementation, and updated the `ApplicationManifestKindValueProvider` to make use of it.