Save property values
Temporarily save certain property values when:
- toggling between icons/manifests and resource files
- toggling documentation file generation
- toggling between package license files and package license expressions

The idea here is that these operations should be non-destructive. That is, if you toggle documentation file generation off, and then back on, we should also restore the provided doc file path (if any). Similarly, if you switch from a package license expression to a file and then back, we should restore the license expression.

We accomplish this by saving the values of the properties to `ITemporaryPropertyStorage` before deleting them, and restore them if/when they later become applicable again.