Replace a tuple with PropertyPageProviderState
Currently, when we create a Project Query API entity representing a property page (a `PropertyPageValue`) we communicate the important state to the producers of the child entities using a tuple. As we add more producers and further refactor this code it has become clear that the use of a tuple here will lead to errors as different producers may accidentally come to expect different tuple shapes. Remembering the correct shape often means going back to the point where the tuple is created, if you can find it.

This commit introduces a new `PropertyPageProviderState` to replace this one particular use of a tuple.