Only set config-independent properties once
Currently, when setting a property value through the Project Query API we first find all the applicable `ConfiguredProject`s, then bind the requested rule in each, and finally retrieve and set the property in each bound rule.

This makes sense for configuration-dependent properties, where we may want to set the values in some configurations but not others. However, most of our properties are configuration-independent and it doesn't matter which `ConfiguredProject` we go through or how many: ultimately, we'll end up with just one definition in the project file, with no configuration conditions. To avoid unnecessary work we should only set the value of a configuration-independent property once.

This commit adds the necessary checks before we start retrieving the `ConfiguredProject`s.