Update unit tests
Update the unit tests around the project action providers. Specifically, the unit tests need to reflect that the updated code path now checks the propety page catalog, and retrieves `Rules` and their `BaseProperty`s in addition to checking `IRule`s and `IProperty`s. Unfortunately this means that we need to set up multiple mocks for what would really be the same data.