Fix file rename crashing when a file changes its extension.
RenameAsync() was trying to apply symbols rename and Roslyn was failing to
find the original filename.

This change checks for changes in file extensions, and avoid calling symbol rename.
This will remove the unnecessary calls to Roslyn.

Apart from this, there are other few changes as part of the cleanup.