Ensure XLF strings do not contain multiple accelerator mnemonics
The ampersand character in UI control labels generally causes the subsequent
character to be treated as a keyboard accelerator.

It is irregular for a UI element to have more than one such accelerator key.

We saw such a situation in a resource file, which was addressed in #6788.

This test ensures the scenario doesn't occur again in future.

Note that not all resource strings are used as UI labels, so it's possible this
test may need tweaking in future.