Source all unresolved references from evaluation
Fixes #6803. Fixes #6804.

The dependencies tree consumes both unresolved and resolved versions of project dependency items.

Unresolved items are obtained quickly via evaluation in order to build the tree rapidly when a project is opened, and to quickly reflect changes to the tree as items are added, removed and updated over time.

Resolved items are obtained via design-time builds. They contain much richer information about dependencies than evaluation items, and the tree is updated with that information as it becomes available.

It's important that unresolved items are only sourced from evaluation. If their rule files request they come from design-time builds, then they the dependencies tree cannot be built until a design-time build completes.

Before this change, both SDK and Analyzer items were being sourced from design-time build targets. This meant that all other dependency types were being held up, waiting for the design-time build to complete. To the user, this means a delay in the construction of the dependencies tree when loading a project.

This change adds unit tests to validate the details of both resolved and unresolved dependency item data sources, to prevent this problem occurring again in future.