Change the default value for Symbols Rename in a fresh VS installation.
In a fresh VS installation or repair, Dotnet ProjectSystem cannot get the value in the checkbox Tools | Options | Project and Solutions | Enable Symbol Rename. At this point VS has no knowledge of this value and that is how the system works. The user has to turn the checkbox off and then on to make it work and have a persisted value for future usages.

This small change sets the default value of EnableSymbolRename to true to match the value on a fresh VS Installation.

This change will enable Symbols Rename by default on a fresh installation.