Add IActiveConfigurationComponent
Similar to IImplicitlyActiveConfigurationComponent, this type is intended for configured-project scope exports to allow them to be activated and deactivated in response to their configuration becoming active or inactive.

The existing IImplicitlyActiveConfigurationComponent does this for _implicitly_ active configurations, of which there may be many (when multi-targeting, for example). This alternative exists for the _singular_ active configuration, which is usually the first when multi-targeting.

This pattern can be useful for deferring initialisation of components until their configuration is activated, reducing work on solution load for example. It can also help clean up components which are no longer applicable when the configuration is changed.