Activate and deactivate the up-to-date check
#6822 made the up-to-date check's initialization lazy, such that it deferred its dataflow subscriptions until its first call. The motivation for that change was to prevent dataflow subscriptions for inactive configurations. Exports for inactive configurations are unused, so initializing them is wasteful and negatively impacts solution load and retained memory usage.

This change had the negative consequence of deferring dataflow subscriptions until the first build, at which time a design-time build would be scheduled to obtain the needed information.

A better solution is to monitor the active project configuration and only activate the up-to-date check component associated with the active configuration. That is what this commit does.

It does so by replacing IProjectDynamicLoadComponent with IActiveConfigurationComponent.