Fix nullable annotations
Currently, the `ILaunchSettingsProvider` interface claims that the `CurrentSnapshot` property and `WaitForFirstSnapshot` method will always return a non-null `ILaunchSettings`. Unfortunately, neither is true.

In practice, creating an `ILaunchSettings` snapshot requires knowing the active profile name, and this is stored in a property in the .user file (if it exists). Retrieving the property requires an evaluation; as such we can't produce an `ILaunchSettings` until at least one evaluation has occurred and we have a chance to process it.

It is possible for a consumer of `ILaunchSettingsProvider` to request the `CurrentSnapshot` before this has occurred, in which case they will receive `null`. Note that the implementation of this property (in the `LaunchSettingsProvider` class) calls `EnsureInitalized()`; this may give the appearance of guaranteeing that the snapshot is available, but all it does is ensure that we are signed up to receive, eventually, the data we need to produce that snapshot.

`WaitForFirstSnapshot` _could_ guarantee that a non-null `ILaunchSettings`--except that the caller may pass a time out, and if the time out expires before a snapshot is available, the call will return `null`.

This change updates the interface to reflect reality, and updates all callers in our code to respond accordingly. In most cases we were already handling the possibility of `null`.