Allow language service to identify dependency tree target nodes
Roslyn needs a way to identify which workspace project context a given dependencies tree node relates to. Ideally we would store the project context ID on the node itself for Roslyn to query, however we are late in the 16.9 cycle and the magnitude of that change would likely make it too risky. For now, we pass the `TargetFramework` property value (aliased as `TemporaryDependencyNodeTargetIdentifier`) to the language service, so it can match the value against the `$TFM:net5.0`-style value found in the node's capabilities property value.