Implement project-level properties
Implement launch profiles "support" for project-level properties. There are no project-level properties supported on the launch profiles; instead the properties all belong to the individual launch profiles. As such, the `IProjectProperties` implementation returned here is largely a bunch of no-ops and exception throws for operations that don't make sense at the project level.

Note that the launch settings *do* have the concept of global settings. However, we currently expose these through the individual launch profiles, we will continue to do so with the `IProjectProperties`.